# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: JINGTANG Platform API
  version: 1.0.0
  description: D2 identity and Workspace boundary plus D4 Source Asset, content revision, and approval operations.
servers:
  - url: /api/v1
paths:
...
  /channels/facebook/data-deletion/status:
    get:
      operationId: readFacebookDataDeletionStatus
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string, pattern: "^META-[A-F0-9]{24}$" }
      responses:
        "200": { description: Opaque deletion status. }
        "404": { description: Confirmation code is absent. }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-jt_session
...
security:
  - sessionCookie: []